Accept 1200
1. Sort the input
2. Push back the data at the middle position of sorted data to the
answer.

If there is a sorted array.
A B C D E
the difference between B and C is less than B and D trivially.
So the answer is C D B E A, just take the middle position of sorted data
and erase it and take ... until there are no data.