Accept 1100
The problem is given n pairs of integers and asking to find n - 1 pairs
of integers that the sum of every pair are not relatively prime.
---
Input: integer n means the number of pair and next 2 * n integer means
the number you should choose to be a pair.
---
1. Just print the index of the pair of odd numbers or even numbers.
Because the number is either odd or even, so is can find the pair either
even or odd at least n - 1 pairs.