Accept 1000
It is a really easy problem
But I did it in complicated way first(I want solve it by just travers
one time).
---
Input:integer n which divisible by 3 and the n integers array
---
Just use a while to check if the a[0-2(means the number of remainders 0,
1, 2)] is zero or not.
If it is less than 0 add the absolute of that to answer and add that number to
next remainder and change it to 0.