Accept 1300
It is a hard math problem
I only sure about that the max of answer is no more than n
There is a proof of why the answer must smaller than n in codeforce
And why the answer is like that
---
Input:a big number n
---
Find the first number x which larger than 1 and n is divisible by it.
and the answer is n / x and n - n / x
And the range of answer is no more than 1000000.
If there is no number x than the answer is 1 and n - 1