Accept 1200
It's an easy problem, but have to set the edge carefully.
---
Input:two integer n and x means the number of elements in array and the
number that have to find the sum of subarray isn't divisible
respictively. And follow n numbers is the elements of array.
---
1. Let s equal the sum of array
2. If the number s isn't divisible by x, answer is n.
3. If the number s is divisible by x, find the first number is not
divisible by x and the last number is not divisible by x.
4. The answer is the maximum of (n - i - 1, j)