Accept 1400
I thouhght is was too complicated at first.In fact the problem is pretty
easy.
----
Input:two integer n and k to indicate the length of initial array and
the length of subarray, and then n integer is the initial array.
----
1. Check if the number of different integer is larger than k, if so,
print -1 and run next test case.
2. So now we know that the number of different integer is not larger
than k, so we just print n * k and then n * k numbers(so we don't have to
calculate how much number should we print)
If the number of different integer is less then k, we can fill the array
we want by 1.