Accept 1300
It's really straight forward to solve it.But I stuck in the edge test
for a long time.
----
Input: Two integers n and k to indicate that the length to string and
the minimum distance between two 1s respectively. And a string construct
by 1 and 0.
----
There are three situations of continuous 0
1. At the first of string and there is 1 after that.
2. At the end of string and there is 1 before that.
3. At the middle of string and there are 1 before and after that.

So, at different situations have to do different actions
1. delete the last k zeros
2. delete the first k zeros
3. delete the first and last k zeros respectively

After all the contunuous 0 are free, you can put 1 in every position,
and the maximum 1 we can put in a block is floor((len(block)+k) / k+1)
The answer is the sum of every blocks' number of 1 that we can put in.