Accept 1200
I went to the wrong way again!
Solve the problem by dfs first and get TLE.
But this problem doesn't need DFS, because there are no any chess on the
board so it is impossible to get in the situation that no way to suit
the board.
It can easily assumning that the first position is B then W then...
And only change the character that the input is "."
----
Input: Two integer to indicate the size of chess board, and the state of
every postion on the chess board
----
For every row in input state
1. Check if the state of that postion is "." if yes goto step 2
2. If the row + col is even then change the content in string to "B"
other wise change it to "W"
3. Print the string