Accept 1200
It's an easy problem and I go the wrong way again.
----
Input: Three integers n, x and y to indicate the number of element in
array and two specific elements should in the array respectively.
And the problem is asking an array that maximum element is  minimal.
----
Just simply trace every factor of y - x (include 1 and (y - x))
Check the factor in ascending if that (y - x) / factor is less than or
equal to n.
The first element satisify the checker is the factor what we want, and
then create the array from y and y - factor and so on, untill enough
element or is not bigger than zero.
If the element is not enough, just append the y + factor and y + 2 *
factor and so on , untill it's enough.