Accept 1500
It's a problem about the maximum difference between the maximal number
and minimal number in multiple decreasing sequence
----
Input: an integer n to indicate the length of array and an array a
----
Traverse the the array and do operation bellow:
1. If the number is less than previous one or less than the first
element in decreasing sequence, check if it is the minimal number in
that sequence.
2. If the number is larger than or equal to previous one or first
element in decreasing means that it is belong to a new decreasing
sequence so check if the distance between minimal and maximal number in
previous sequence is larger than answer.
After the traverse we can get the largest distance, and than find the
minimum x-power of 2 larger than the distance.
The x is the answer!