Accept 1400
Traverse and check the number of change have to do and save the minimal
step of change
----
Input: a string s
----
1. Count the number of 1 and 0 in s.(num_one, num_zero)
2. Traverse the whole string and maintain the number of 1 and 0 that
have traverse.(tra_one, tra_zero)
3. In every index X you can calculate the cost that from index 0 to X
is 1 and the other is 0 and from 0 to X is 0 and the other is 1.
	cost(11...00, X) = tra_zero + num_one - tra_one
	because the zero that traverse have to change to 1 and the
	remain 1s have to change to 0, and so do cost(00....11, X)
So you only have to keep the minimal cost after traverse and that is the
answer.