Accept 900
Use the tree class at first and can't solve problem correctly.
So change the way to solve it.
The main idea is DFS to find the maxmimal depth between multiple trees.
----
Input: an integer n to indicate the number of employees and an array A
to indicate the manager index of every employee i, if index is -1 means
that the i-th employee does not have an immediate manager
----
1. Transform the array to an vector called root and an array called
manager that indicate the i-th employee doesn't have manager and the
employee of i-th manager respectively.
2. Use DFS to find the maximal depth of manager-employee path from a
root.
DFS(int employee, int dep){
	depth[employee] = dep;
	ans = max(ans, dep);
	for every employee x that employee manage by
		DFS(x, dep + 1)
}
3. Do DFS to every employees in root to find the maximal depth.