Accept 1100
I think this problem is easy and I wanna practice the tree problem but
this problem doesn't have to use trees
----
Input: two integer n and q to indicate the length of two arrays
respectively, and two array a and p to indicate the color array have to
find and a question array
----
1. Use map to record the index when a new color appearance and the index
of map is the color number.
2. Use map to find the index of question array asking.
3. After finding the index x, add 1 to the index number before x
recorded in the map, and then change x to zero.