Accept 1200
Wanna solve it by map first but it may be a wrong way, so I solve it by
vector.
---
Input: an integer n to indicate the number of nodes and the array a have
n - 1 elements to indicate the 2 ~ n nodes' parent
----
1. Maintain two vector leaves and child, If the new parent is a node's
leaves then remove it from leaves vector.
2. Check every vector if there are child in vector, if there is a node's
leaves size is less than 3 and have child then the tree is not spruce